requirements.txt
bs4
lxml
requests-html
